feat: Add options for configuring inlay hints (#98)
Just needed to bump the minimum vscode version as 1.66 is the first version to support inlay_hints in the LSP spec.

Also removes the need to pass options to lsp via command line args. Instead, we now call which passes along all the config options as a json type that gets deserialized in the server on initialization.

initializationOptions: workspace.getConfiguration("sway-lsp")